revert test_event_collection
